backend/requirements.txt
# AKA-Sim Backend Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-socketio[asgi]>=5.9.0

# PyTorch
torch>=2.0.0
torchvision>=0.15.0

# Data
numpy>=1.24.0
pydantic>=2.0.0
pandas>=2.0.0
pyarrow>=12.0.0
pillow>=9.0.0

# HuggingFace
huggingface_hub>=0.16.0

# Development
httpx>=0.24.0

python-multipart
einops